test(e2e): keep the interface-save wait in one place
The theme scenario's new wait and the language helper's own were the same
request-bound wait on the same endpoint, written twice. Two copies drift, and a
fix applied to either would have missed the other — the shape this campaign is
otherwise busy closing inside the Go tests.

Both now call saveInterfaceSettingsForm. The success-flash readback is opt-in:
it is the only server-backed half the theme has, since the theme itself lives in
client storage by design, but the language callers leave /settings straight away
and never land on the page that renders it. Their behaviour is unchanged, and
saveSettingsLanguage keeps its signature, so the four specs importing it are
untouched.
