+++ e2e/support/language-helpers.ts
@@ -1,5 +1,7 @@
 import { expect, type Page } from '@playwright/test';
 
+import { saveInterfaceSettingsForm } from './settings-interface-helpers';
+
 export async function switchPublicLanguage(page: Page, code: string): Promise<void> {
   const form = page.locator('[data-language-switch-form]');
   const button = form.locator(`[data-language-switch-option="${code}"]`);
@@ -26,19 +28,12 @@ export async function saveSettingsLanguage(page: Page, code: string): Promise<vo
     await option.locator('.chip-stack').click();
     // The radio change flips data-selected optimistically on the client, so the
     // assertion below can pass before the htmx PATCH persists the preference.
-    // Bind to the save click's own request and await its response so callers
-    // that navigate away next (e.g. straight to /calendar) don't race an
-    // in-flight save that would otherwise drop the just-chosen language.
-    const [saveRequest] = await Promise.all([
-      page.waitForRequest(
-        (request) =>
-          request.method() === 'PATCH' &&
-          request.url().includes('/api/v1/users/current/interface'),
-      ),
-      form.locator('[data-settings-interface-save]').click(),
-    ]);
-    const saveResponse = await saveRequest.response();
-    expect(saveResponse?.ok()).toBeTruthy();
+    // The shared helper waits on the save click's own request and its response,
+    // so callers that navigate away next (e.g. straight to /calendar) don't race
+    // an in-flight save that would otherwise drop the just-chosen language.
+    // The success flash is not asserted here: those callers leave /settings
+    // immediately and never land on the page that renders it.
+    await saveInterfaceSettingsForm(page);
   }
 
   await expect(option).toHaveAttribute('data-selected', 'true');
